+++ test/test-css-files.php
@@ -19,4 +19,4 @@ $_oGenerator = new PHPClassMapGenerator(
     ]
 );
 
-print_r( $_oGenerator->getItems() );
+print_r( $_oGenerator->get() );